requirements.txt
polars
pyecharts
QuickEcharts
nicegui-tabulator
plotly
great_tables